improve github handler
this handler is now taking into account the current extension of the
distfile as well as explicit assets associated with a release. this
prevents marking something like 'polybar-3.4.3.tar' as
a newer release than 'polybar-3.4.3.tar.gz'.

it it won't try to mark 'release-1.2.3' as a difference to '1.2.3'.
this fixes aria2 at least.